+++ src/test/ft_strncmp.c
@@ -1,4 +1,6 @@
-int	ft_strncmp(const char *s1, const char *s2, int n) {
+#include <stddef.h>
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n) {
 	unsigned int	i;
 	int				aux;
